Fix 20 bugs found in comprehensive codebase audit
CRITICAL fixes:
- xnodeconnector.cpp: bitwise OR (|) -> logical OR (||) in popup check
- tds.cpp: array underflow yin[ch_cnt-1] when ch_cnt=0, reorder NR_P before YMU/YOF
- xpythonsupport.cpp: GIL released during Python object operations in kame_mainthread
- pythondriver.h: re-enable GIL acquire in Payload::local()
- xpythonsupport.cpp: initialize *ret=py::none() before try block in mainthread_callback
- secondarydriverinterface.h: bounds-check std::find() result before dereference
- allocator_prv.h: add virtual destructor to PoolAllocatorBase

HIGH fixes:
- analyzer.cpp: null-check m_entries.lock() in XValGraph and XCalibratedEntry
- analyzer.cpp: capture m_entry before loop to avoid race with onSelectionChanged
- cyfxusb_libusb.cpp: safe bounds for USB string descriptor buffer
- pythondriver.h: null-check before pushing to garbage queue
- xpythonsupport.py: fix thread ID comparison using native_id
- xnodeconnector.cpp: fix uninitialized m_pWindow when window=nullptr

MEDIUM fixes:
- lecroy.cpp: cast to size_t to prevent integer overflow in buffer calc
- xpythonsupport.py: skip comment lines in .kam Ruby->Python translation
- xnodeconnector.cpp: disconnect before connect in QColorDialog to avoid duplication
- tds.cpp: increase snprintf buffer from 9 to 16 bytes
- xpythonmodule.cpp: catch pybind11::cast_error in payload down-caster
- CMakeLists.txt: fix test name typo (negotioation -> negotiation)

LOW fix:
- xnodeconnector.cpp: NULL -> nullptr in disconnect() call

Includes KAME_BugCheck_Report.docx with full details.

+++ modules/dso/lecroy.cpp
@@ -420,8 +420,8 @@ XLecroyDSO::convertRaw(RawDataReader &reader, Transaction &tr) {
 		fprintf(stderr, "first_valid=%d,last_valid=%d,first=%d,acqcount=%d,count=%d\n",
 			(int)first_valid, (int)last_valid, (int)first, (int)acqcount, (int)count);
 		if(ch == 0) {
-			if((count < 0) || 
-				(reader.size() < (count * 2 + DATA_BLOCK + n + 2) * ch_cnt))
+			if((count < 0) ||
+				(reader.size() < ((size_t)count * 2 + DATA_BLOCK + n + 2) * ch_cnt))
 				throw XBufferUnderflowRecordError(__FILE__, __LINE__);
 			tr[ *this].setParameters(ch_cnt, hoffset, interval, count);
 		}
